requirements.txt
kafka-python==2.0.2
pandas==2.2.0
openpyxl==3.1.2
openai==1.12.0
python-dotenv==1.0.0
requests==2.31.0